remove flask
