+++ pyproject.toml
@@ -12,9 +12,7 @@ psycopg2 = "^2.9.9"
 openai = "^1.33.0"
 together = "^1.2.1"
 ollama = "^0.2.1"
-flask = "^3.0.3"
 pyyaml = "^6.0.1"
-flask-cors = "^4.0.1"
 anthropic = "^0.29.0"
 fastapi = "^0.111.1"
 uvicorn = "^0.30.3"
@@ -24,7 +22,6 @@ chat_with_tool = "examples.chat_with_tool:main"
 eval = "eval.run_eval:main"
 chat_server = "examples.chat_server.server:main"
 ddn_setup = "scripts.ddn_setup:main"
-chat_server_streaming = "examples.chat_server.fastapi_server:main"
 
 [tool.poetry.group.dev.dependencies]
 ipython = "^8.25.0"
